backend/recommendplant/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
django-cors-headers==3.14.0
django-environ==0.10.0
django-sslserver==0.22
djangorestframework==3.14.0
google-api-core==2.11.0
google-auth==2.16.3
google-cloud-aiplatform==1.23.0
google-cloud-bigquery==3.8.0
google-cloud-core==2.3.2
google-cloud-resource-manager==1.9.1
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
grpc-google-iam-v1==0.12.6
grpcio==1.53.0
grpcio-status==1.53.0
idna==3.4
mysql==0.0.3
mysqlclient==2.1.1
numpy==1.24.2
packaging==21.3
pandas==1.5.3
Pillow==9.4.0
proto-plus==1.22.2
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.2
requests==2.28.2
rsa==4.9
Shapely==1.8.5.post1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.15
